mave each panel to its own file
