+++ server/settings.json
@@ -1,1 +1,1 @@
-"{\"temp_enabled\":false,\"temp_sensitivity\":42.5,\"accel_enabled\":true,\"accel_sensitivity\":57.8,\"refresh_time\":1.0}"
+"{\"temp_enabled\":false,\"temp_sensitivity\":42.5,\"accel_enabled\":true,\"accel_sensitivity\":57.8,\"refresh_time\":5.0}"
